requirements.txt
alembic==1.13.2
aniso8601==9.0.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
Babel==2.15.0
black==24.8.0
bleach==6.1.0
blinker==1.8.2
cffi==1.16.0
click==8.1.7
coverage==7.6.1
cryptography==43.0.0
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
execnet==2.1.1
Faker==26.1.0
Flask==3.0.3
Flask-Admin==1.6.1
Flask-BabelEx==0.9.4
Flask-Ext==0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Mailman==1.1.1
Flask-Migrate==4.0.7
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.5.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
importlib_resources==6.4.0
infinity==1.5
iniconfig==2.0.0
intervals==0.9.2
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pur==7.3.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pytest==8.3.2
pytest-cov==5.0.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ldap==3.4.4
pytz==2024.1
six==1.16.0
speaklater==1.3
SQLAlchemy==2.0.32
SQLAlchemy-Utils==0.41.2
tomli==2.0.1
typing_extensions==4.12.2
validators==0.33.0
watchdog==4.0.1
webencodings==0.5.1
Werkzeug==3.0.3
WTForms==3.1.2
WTForms-Alchemy==0.18.0
WTForms-Components==0.10.5
